Structure param deserializers some more
